Change EL Parser to Oracle EL to fix problem with null as empty strings.
In EL 3.0 specification a relevant change was introduced relative to how
a null value must be treated. The specification considers that an empty 
string should be used which may be considered a specification failure.
The Apache Parser faithfully implements the specification while the 
Glassfish (Oracle) does not. 

For more info please refer to Bauke Scholtz's article "The empty String 
madness": 
http://balusc.omnifaces.org/2015/10/the-empty-string-madness.html. 

Only the EL parsser library changes. For 
compatibility reasons submitted values ​​may not be interpreted as null.
This is the reason why context-param 
javax.faces.INTERPRET_EMPTY_STRING_SUBMITTD_VALUES_AS_NULL is forced to
false, unlike what is suggested.