Improve workflow form nodes of type "SelectMenuForm".
The following properties have been added:

- Layout: How the form will be displayed. There are three possible 
values:
  - List: Vertical list of options.
  - Assistant: Horizontal list of options. In this layout, a div with 
class = 'bot' will be added to the form to simplify the CSS definition 
of a virtual assistant if necessary.
  - From CSS file: The menu will be displayed with a layout described 
in the file whose URL is specified in the "CSS File URL" property.

- CSS File URL: The URL of a CSS file that will be linked directly 
within the page when the node is executed. It can be a full URL 
(starting with 'http' or 'https') or a local URL to the execution 
context.

- Custom CSS: CSS code that will be inserted directly into the page 
when the node is executed.