Allow the display of news that redirect to customized links.
A separation mark may be included in the headline of a news item to 
separate the headline into two parts: the actual headline and a custom 
URL. The actual headline will be displayed on the news pages, and if a 
link is displayed it will redirect the user to the specified URL.

The separation mark can be configured with the node property 
"newsURLSeparator". By default it will be "###".