Improve the custom links system in the news module.
The new 'customUrl' variable has been added to the objects of type 
'New' and 'NewView' to store the custom destination URL to be used when 
clicking on the news item. The new 'CUSTOMURL' column must be added to 
the 'INF_NEW' table to persist the value of the variable.

This new functionality will work in both the news search pages and the 
'HtmlNews' and 'HtmlNewsCarousel2' widgets.