Allow external links with custom target type in the news module.
The new 'customUrlTarget' variable has been added to the objects of 
type 'New' and 'NewView' to store the target type of the custom 
destination URL link, if exists. The new 'CUSTOMURLTARGET' column must 
be added to the 'INF_NEW' table to persist the value of the variable.

This new functionality will work in both the news search pages and the 
'HtmlNews' and 'HtmlNewsCarousel2' widgets.