Allow preferred image types in news RSS servlet.
The 'img' parameter has been added to the RSS servlet URL to specify 
the preferred type of the displayed image in each entry.

It supports a string of up to 3 chars, which can contain only the 
letters:

- 'l' to show a list image, if it exists.
- 'd' to show a detail image, if it exists.
- 'c' to show a carousel image, if it exists.

Examples:

- "https://<server>/rss/<section>?img=cdl" returns a feed containing 
for each entry an image of type carousel, detail or list, preferably in 
this order.

- "https://<server>/rss/<section>?img=c" returns a feed containing for 
each entry an image of type carousel, not showing any if there is no 
such image.

If the parameter is not specified, the servlet will use the priority 
'ldc'.