New users lock system and other improvements in security module.
New users lock system to allow a user to be locked if a login attempt 
limit has been reached.

The system will automatically unlock a locked user after a certain 
amount of time since the lock.

If a locked user exceeds a certain number of failed login attempts, the 
system will record an intrusion attempt.

The admin and autologin users can never be locked.

The default lock system behavior can be overridden with the next system 
properties (to be defined in web.xml or in root.xml):

- org.santfeliu.security.service.SecurityManager.userLockControlMode: 
users on which the lock system will be applied. Possible values: [all, 
noldap, off] (default: off). 
  - all: All users.
  - noldap: Users not present in the active directory.
  - off: The system is inactive for all users.

- org.santfeliu.security.service.SecurityManager.maxFailedLoginAttempts: 
number of failed login attempts that will cause the user to be locked 
(default: 5).

- org.santfeliu.security.service.SecurityManager.autoUnlockMarginTime: 
seconds that have to pass before a locked user is automatically 
unlocked (default: 600).

- org.santfeliu.security.service.SecurityManager.minIntrusionAttempts: 
number of failed login attempts that will cause a user intrusion 
attempt to be recorded in the database (default: 10).

Other improvements in the security module:

- Improved autologin user security.
- Improved changePassword method.
- The removeUser method will also remove the user properties stored in 
the USER_PREFERENCES table.
- The removeUser method will return true if the removal was successful 
(until now it always returned false).
- The storeUserProperties and removeUserProperties methods can only be 
executed with administrator credentials.