Add and improve URL parameters in news RSS servlet.
The 'text' parameter has been added to the RSS servlet URL to specify 
which text to include in each entry.

It supports a string of up to 2 chars, which can contain only the 
letters:

- 's' to show the summary.
- 't' to show the full text.

Examples:

- "https://<server>/rss/<section>?text=t" returns a feed containing for 
each entry only the full text, not the summary.

- "https://<server>/rss/<section>?text=st" returns a feed containing 
for each entry the summary and the full text.

If the parameter is not specified, the servlet will use the value 's' 
(only the summary will be displayed).

In addition, the value 'all' has been added to the 'img' parameter, to 
show all the images of the news, regardless of their type.