New JSF custom function 'toUuid'.
A new JSF custom function has been added with this signature:

java.lang.String toUuid(java.lang.String)

It receives as a parameter a resource URL that contains a docId and 
returns the same URL with the docId converted to contentId, if the 
current user has privileges to read the document.

Accepted URL format: [/documents/]docId[/filename][?params]

If the parameter does not have this format, or the user does not have 
read access to the document, it returns the same parameter without 
changes.