Improve management of pages with groupable rows.
The GroupableRowsHelper class has been added to unify the code that 
handles row grouping. This new helper will be used in all pages with 
groupable rows:

Pages using <p:columns>:

- case_persons
- case_addresses
- case_interventions
- case_cases
- case_documents
- case_events
- event_cases

Pages using some <p:column> tags:

- event_persons
- event_documents

Pages using some <p:column> tags (row grouping is allowed from now on):

- person_cases
- document_cases
- address_cases
- new_documents

The new "groupBy" property has been added to the edit tabs defined in 
ObjectSetup files. This new property has the next syntax:

"groupBy": "<groupBy>[:<groupByDescription>]" 

- <groupBy> represents the value/expression to be evaluated to assign a 
row to a certain group.
- <groupByDescription> represents the value/expression containing the 
description of the group to which the row is assigned. If it does not 
exist, the value of <groupBy> will be displayed.

Both <groupBy> and <groupByDescription> can be column names or EL 
expressions. If a column name is a type field, the type description 
will be returned.

Examples:

- “groupBy”: ”eventDocTypeId”: the rows will be grouped by the type 
description, and the group description will be the same value.
- "groupBy": "#{row.typeId}:casePersonTypeId": the rows will be grouped 
by the raw typeId field, and the group description will be the type 
description.
- “groupBy”: “p2”: the rows will be grouped by the 'p2' metadata value.

A new "hideColumnWhenGrouping" optional property has been added to the 
edit tabs (within the 'properties' object) to specify which column will 
be hidden when grouping rows (if not specified, it will be the column 
specified in "groupBy").

The "groupedViewEnabled" edit tab property can no longer be used, 
because it will implicitly be "true" when the new "groupBy" property 
has a defined value.

Additionally, the "orderBy" property has been added to the event_cases 
page.