Improve visual form accessibility tools in MatrixIDE.
- New button in the MatrixIDE tool: "Repair info texts". This button 
will perform the following automated tasks:

  - If an input component has an OutputText component immediately to 
its right, the latter will be removed and its text value will be 
assigned to the "infoText" field of the input component.
  - If an input component has an OutputText component immediately below 
it, the latter will be removed and its text value will be assigned to 
the "helpText" field of the input component.
  - FormImporter has been fixed to allow labeling of input components 
with info texts and/or icons.
  - The "helpText" property will be displayed in OutputTextArea 
components.

- Label repair has been improved:

  - If the "forElement" field of a Label component points to a 
non-existent id, it will be deleted and a recalculation will be 
attempted.
  - If an id must be generated for an input component, the same text as 
in the "variable" field will be used, if this is defined (except for 
radio inputs).
  - If a component of type "checkbox" or "radio" has an output 
component immediately to its right with text not ending with ":", this 
will be the component designated as the new label. Otherwise, if 
the component has an output component immediately to its left with text 
ending with ":", this will be the designated component.