Add cache of enum type items.
A new enum type item cache has been added to the platform. It is 
available with the "EnumTypeCache.getInstance()" method, and contains 
the following methods:

- getItems(enumTypeId): Returns a list of items of the specified type, 
or null if the type does not exist.

- getItemsByValue(enumTypeId, value): Returns a list of items of the 
specified type whose value is equal to the specified one, or null if 
the type does not exist.

- getItem(enumTypeItemId): Returns the item with the specified id, or 
null if the item does not exist. If the type to which the element 
belongs was not cached, the type is added to the cache.

- clear(): Removes all cache entries.

- clear(enumTypeId): Removes the specified type in the cache.

The new cache has been used in the 'dic', 'query' and 'report' module 
pages that made web service calls to work with enum types and items, 
and in the DataTableRow class.