Allow filtering by custom properties in tab pages.
The 'TypeSelectHelper' class has been replaced by the new 
'RowsFilterHelper' class.

This class allows filtering by a custom property specified in the new 
'filterBy' property, to be used on edit tabs specified in ObjectSetup 
files.

The format of this property is...

"filterBy": "<property_name>[:<label>]"

...where:

- <property_name> is the name of the property to filter by. This can 
be a native row field or a named custom property.
- <label> is the label for the filter selector. Default: "Filter".

If the "groupBy" property is specified, the "filterBy" property will 
be ignored.

If "groupBy" and "filterBy" are not specified, the page will be 
displayed filtered by the row type.

In all cases, the selector will only be displayed if there is more 
than one option available.

This new feature has been added to the case tabs and to the 
'address_cases', 'person_cases' and 'person_events' tabs.