Add a new action to copy an object.
The new "copyObject" action has been added to the predefined actions 
for an object. This action makes a copy of the current object, 
applying any unsaved changes.

For this action to appear in the object action list, the object type 
must have the "copyEnabled"="true" property.

The "preCopy" and "postCopy" script actions have been added.