Improve volume management in the case_documents page.
- In the volume selector, the empty option will now select documents 
without a volume, not all documents.
- An optional "Show All" option has been added to the volume selector. 
To see this option, the "enableShowAllVolumes" property must be set to 
'true' in the corresponding edit tab of the ObjectSetup file.
- When loading the tab, a default volume can be displayed if the 
"volume" property is defined in the corresponding edit tab of the 
ObjectSetup file. Use "UNDEFINED_VOLUME" to view documents without a 
volume.
- The "volume" column is now visible if it is defined in the 
ObjectSetup file. This column will only be displayed when the 
"Show All" option is selected.
- The number of rows displayed has been increased to 10.
- Volumes in the selector will be displayed in alphabetical, ascending 
order.