Add sort icons to the column headers in the document_list page.
New icons have been added to the column headers displayed in the 
document_list page to sort the list by the selected column value. 
These icons allow the user to switch between ascending and descending 
order.

The columns with these icons can be specified in the "orderByColumns" 
property of SearchTab objects in ObjectSetup files. The property must 
contain an array of column names. The default value is ["docId", 
"title", "docTypeId"].