Add a description of the query execution to the 'sqlweb' page.
The 'sqlweb' page edit panel will display:

- The SQL code, if the user can edit queries.
- A description of the query execution, otherwise.

In both cases, the edit panel will be hidden by default. It can be 
made visible with the eye icon.

This icon will only be displayed if the 'sqlweb' page is opened from 
the 'query_view' page. If the user accesses the 'sqlweb' page 
directly, this icon will be hidden, and the edit panel will always be 
visible.