Allow a second linked row filter in object tabs.
The "filterBy" property of EditTabs in ObjectSetup files can now 
specify a second level of filtering, depending on the value chosen in 
the first filter.
 
This will be the format of the property from now on:

"filterBy": {"name": "p1", "label": "Prop 1",
  "linkedFilterBy": {
    "A": { "name": "p11", "label": "Prop 11" },
    "B": { "name": "p12", "label": "Prop 12" },
    "C": { "name": "p13", "label": "Prop 13" },
    ...
    "N": { "name": "pN", "label": "Prop N" }
  }
}

In the example:

- If the value of property "p1" is "A", then a second filter by 
property "p11" will be displayed, with the label "Prop 11".
- If the value of property "p1" is "B", then a second filter by 
property "p12" will be displayed, labeled "Prop 12".
- And so on...

For the value (A, B, C, ...) of "linkedFilterBy", the characters "*" 
or "%" can be used as wildcards. In this case, regardless of the value 
chosen in the first filter, the second filter will be displayed.