Allow modification of search and edit tab labels.
- New multivalued property "searchTabLabel" for CMS nodes to specify 
the label for each search tab. If it doesn't exist, the value of the 
"label" property of each SearchTab in the ObjectSetup file will be 
used. If that doesn't exist, the value of [objectBundle.list] will be 
used.

- New property "editTabLabel" for CMS nodes to specify the edit tab 
label. If it doesn't exist, the value of [objectBundle.edition] will 
be used.