Improve the case_documents tab.
- A TabView component with three tabs has been added to the edit 
dialog of a CaseDocument object:
  - "Create new document": Allows the user to select a document from 
the device and upload it to the document manager as a new document, 
linking it to the file.
  - "Existing document": Allows the user to select a previously 
created document from the document manager and link it to the file.
  - "Update content": Allows the user to select a document from the 
device to replace the content of the linked document, creating a new 
version. This tab only appears when editing a CaseDocument, not when 
creating a new one.
- When a file is selected in any of the tabs, the others are 
deactivated, to avoid conflicts.
- Added the new MatrixConfig property 
"org.santfeliu.doc.fileUploadMaxSize" to indicate the maximum size of 
the files to upload, in bytes. Default value: 200000000 (200 MB).
- Removed "Upload" and "Replace Content" buttons that invoked the 
MatrixClient application.
- Added "Type" field to the document_cases tab. Some labels and error 
messages have also been improved.